Restructure application
