+++ Dockerfile
@@ -5,4 +5,4 @@ RUN npm ci
 COPY . .
 RUN npm run build
 EXPOSE 4000
-CMD ["node", "server.mjs"]
+CMD ["node", "dist/server.mjs"]
