+++ codegen.yml
@@ -1,5 +1,5 @@
 overwrite: true
-schema: "schema.graphql"
+schema: "src/schema.graphql"
 documents: null
 generates:
   src/schema.ts:
