+++ package.json
@@ -2,12 +2,12 @@
   "name": "sul-folio-graphql-server",
   "version": "1.0.0",
   "description": "",
-  "main": "server.mjs",
+  "main": "dist/server.mjs",
   "type": "module",
   "scripts": {
     "build": "npm run codegen && npm run compile",
     "compile": "tsc",
-    "start": "npm run codegen && ts-node-esm src/index.ts",
+    "start": "npm run codegen && ts-node-esm src/server.mjs",
     "start:dev": "nodemon",
     "codegen": "gql-gen --config codegen.yml",
     "lint": "eslint src -c .eslintrc.yml",
