+++ src/folio/circulation-api.ts
@@ -1,5 +1,5 @@
 import FolioAPI from "./folio-api.js"
-import { CirculationLoan } from './schema'
+import { CirculationLoan } from '../schema'
 
 export default class CirculationAPI extends FolioAPI {
   async getLoan(id: string): Promise<CirculationLoan> {
