+++ src/folio/feefines-api.ts
@@ -1,5 +1,5 @@
 import FolioAPI from "./folio-api.js"
-import { FeeFine } from './schema'
+import { FeeFine } from '../schema'
 
 export default class FeeFinesAPI extends FolioAPI {
   async getFeeFine(id: string): Promise<FeeFine> {
