+++ src/folio/locations-api.ts
@@ -1,5 +1,5 @@
 import FolioAPI from "./folio-api.js"
-import { ServicePoint } from './schema'
+import { ServicePoint } from '../schema'
 
 export default class LocationsAPI extends FolioAPI {
   async getLocation(id: string): Promise<ServicePoint> {
