+++ src/server.mjs
@@ -6,25 +6,30 @@ import express from 'express';
 import http from 'http';
 import cors from 'cors';
 import bodyParser from 'body-parser';
-import { resolvers } from './dist/index.js';
 
 import { readFileSync } from "fs"
+import path from 'path';
+import url from 'url';
+const __filename = url.fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 import Honeybadger from '@honeybadger-io/js';
 
-import PatronsAPI from "./dist/patrons-api.js"
-import UsersAPI from "./dist/users-api.js"
-import LocationsAPI from "./dist/locations-api.js"
-import InstancesAPI from "./dist/instances-api.js"
-import ItemsAPI from "./dist/items-api.js"
-import HoldingsAPI from "./dist/holdings-api.js"
-import TypeAPI from "./dist/type-api.js"
-import FeeFinesAPI from "./dist/feefines-api.js"
-import FolioAPI from "./dist/folio-api.js"
-import CirculationAPI from "./dist/circulation-api.js"
-import OkapiAPI from "./dist/okapi-api.js"
+import { resolvers } from './folio/index.js';
+import PatronsAPI from "./folio/patrons-api.js"
+import UsersAPI from "./folio/users-api.js"
+import LocationsAPI from "./folio/locations-api.js"
+import InstancesAPI from "./folio/instances-api.js"
+import ItemsAPI from "./folio/items-api.js"
+import HoldingsAPI from "./folio/holdings-api.js"
+import TypeAPI from "./folio/type-api.js"
+import FeeFinesAPI from "./folio/feefines-api.js"
+import FolioAPI from "./folio/folio-api.js"
+import CirculationAPI from "./folio/circulation-api.js"
+import OkapiAPI from "./folio/okapi-api.js"
 
 // Read the schema.graphql into utf-8 string so we can pass it to Apollo
-const typeDefs = readFileSync("schema.graphql").toString("utf-8")
+const typeDefs = readFileSync(path.resolve(__dirname, "schema.graphql")).toString("utf-8")
 
 class AuthnAPI extends FolioAPI {
